Split DagOps to several files, to be compiling yet.
