+++ ailets-rs/gpt/src/lib.rs
@@ -4,7 +4,7 @@ pub mod handlers;
 pub mod structure_builder;
 
 use actor_io::{AReader, AWriter};
-use dagops::{DagOpsTrait, DummyDagOps};
+use dagops::{InjectDagOpsTrait, InjectDagOps};
 use handlers::{
     on_begin_message, on_choices, on_content, on_end_message, on_function_arguments,
     on_function_begin, on_function_id, on_function_index, on_function_name, on_role,
@@ -142,7 +142,7 @@ fn make_triggers<'a, W: Write + 'a>() -> Vec<Trigger<'a, BA<'a, W>>> {
 pub fn _process_gpt<W: Write>(
     mut reader: impl std::io::Read,
     writer: W,
-    dagops: &impl DagOpsTrait,
+    dagops: &impl InjectDagOpsTrait,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let builder = StructureBuilder::new(writer);
     let builder_cell = RefCell::new(builder);
@@ -185,7 +185,7 @@ pub extern "C" fn process_gpt() {
     let writer = AWriter::new(c"").unwrap_or_else(|e| {
         panic!("Failed to create writer: {e:?}");
     });
-    let dagops = DummyDagOps::new();
+    let dagops = InjectDagOps::new();
     _process_gpt(reader, writer, &dagops).unwrap_or_else(|e| {
         panic!("Failed to process GPT: {e:?}");
     });
